+++ lib/src/utilities/utilities.dart
@@ -1,0 +1,8 @@
+library utilities;
+
+import 'package:layrz_models/layrz_models.dart';
+
+// Duration humanizer
+part 'src/duration/config.dart';
+part 'src/duration/enums.dart';
+part 'src/duration/extension.dart';